requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.8.1
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
black==23.12.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.8.0
exceptiongroup==1.1.3
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
huggingface-hub==0.17.3
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
openai==1.1.2
OpenTimelineIO==0.15.0
orjson==3.9.10
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
psycopg2-binary==2.9.9
py-cord-dev==2.5.0rc5
pyaaf2==1.6.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
pytube==15.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlalchemy-cockroachdb==2.0.1
tokenizers==0.14.1
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
yarl==1.9.2